requirements.txt
aiogram==2.22.1
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.6.15.1
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1
django-on-heroku==1.1.2
frozenlist==1.3.1
gunicorn==20.1.0
idna==3.3
multidict==6.0.2
psycopg2-binary==2.9.3
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
yarl==1.8.1
